Stage-2/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
pydantic==2.12.3
python-dotenv==1.2.1
pillow==12.0.0
requests==2.32.5
CairoSVG==2.8.2
